fixing conflicts and updating tests
