+++ tests/test_main_measurements.py
@@ -3,7 +3,10 @@ from __future__ import annotations
 import numpy as np
 import pytest
 
-from cavendish_particle_tracks._main_widget import ParticleTracksWidget
+from cavendish_particle_tracks._main_widget import (
+    IMAGE_LAYER_NAME,
+    ParticleTracksWidget,
+)
 
 
 @pytest.mark.parametrize(
@@ -45,10 +48,9 @@ def test_calculate_radius_ui(
     for expected in expected_lines:
         assert expected in captured.out
 
-    assert cpt_widget.table.item(0, 2)
-    assert cpt_widget.table.item(0, 3)
-    assert cpt_widget.table.item(0, 4)
-    assert cpt_widget.table.item(0, 5)
+    assert cpt_widget.table.item(0, cpt_widget._get_table_column_index("radius_px"))
+    assert cpt_widget.table.item(0, cpt_widget._get_table_column_index("radius_cm"))
+    assert cpt_widget.table.item(0, cpt_widget._get_table_column_index("rpoints"))
 
     assert cpt_widget.data[0].radius_px == pytest.approx(rad, rel=1e-3)
 
@@ -75,10 +77,7 @@ def test_calculate_radius_fails_with_wrong_number_of_points(
     cpt_widget._on_click_radius()
     captured = capsys.readouterr()
 
-    assert (
-        "ERROR: Select three points to calculate the path radius."
-        in captured.out
-    )
+    assert "ERROR: Select three points to calculate the path radius." in captured.out
 
 
 @pytest.mark.parametrize(
@@ -96,9 +95,9 @@ def test_calculate_radius_fails_data_out_of_sync(
     rad,
 ):
     """Test that the radius cannot be computed if the data is out of sync."""
-    # Add images to the viewer
+    # Add images to the viewer (view, event, y, x)
     images = np.random.randint(0, 10, (3, 5, 10, 10), "uint8")
-    cpt_widget.viewer.add_image(images, name="Particle Tracks")
+    cpt_widget.viewer.add_image(images, name=IMAGE_LAYER_NAME)
     cpt_widget.viewer.dims.set_current_step(0, 0)  # move to view 0
     cpt_widget.viewer.dims.set_current_step(1, 0)  # move to event 0
 
@@ -138,15 +137,8 @@ def test_calculate_radius_fails_data_out_of_sync(
     ), "The radius should not be calculated"
 
     assert not cpt_widget.table.item(
-        0, 2
+        0, cpt_widget._get_table_column_index("rpoints")
     ), "The radius points should not be recorded"
-    assert not cpt_widget.table.item(
-        0, 3
-    ), "The radius points should not be recorded"
-    assert not cpt_widget.table.item(
-        0, 4
-    ), "The radius points should not be recorded"
-    assert not cpt_widget.table.item(0, 5), "The radius should not be recorded"
 
     # change the data view so that it's in sync
     cpt_widget.viewer.dims.set_current_step(0, 0)  # move to view 0
@@ -157,15 +149,49 @@ def test_calculate_radius_fails_data_out_of_sync(
     ), "The radius should have been calculated"
 
     assert cpt_widget.table.item(
-        0, 2
-    ), "The radius points should have been recorded"
+        0, cpt_widget._get_table_column_index("radius_px")
+    ), "The radius should have been recorded"
     assert cpt_widget.table.item(
-        0, 3
+        0, cpt_widget._get_table_column_index("rpoints")
     ), "The radius points should have been recorded"
-    assert cpt_widget.table.item(
-        0, 4
-    ), "The radius points should have been recorded"
-    assert cpt_widget.table.item(0, 5), "The radius should have been recorded"
+
+
+def test_radius_save_preserves_old_data(cpt_widget):
+    """Test that previous particles' saved radii are not changed by current particle."""
+    measurements_layer = cpt_widget._setup_measurement_layer()
+
+    # Add first particle
+    cpt_widget.particle_decays_menu.setCurrentIndex(1)  # adds new row
+    # Add three points on a circle of radius 1 px
+    measurements_layer.add([[0, 0, 0, 1], [0, 0, 1, 0], [0, 0, 0, -1]])  # 1px radius
+    # Select the first three points in the layer
+    measurements_layer.selected_data = {0, 1, 2}
+    # User clicks the radius calculation button
+    cpt_widget._on_click_radius()
+
+    # Create a second particle, add points corresponding to a 5 px radius, and click calculate radius.
+    cpt_widget.particle_decays_menu.setCurrentIndex(1)
+    measurements_layer.add([[0, 0, -6, 3], [0, 0, -3, 2], [0, 0, 0, 3]])  # 5px radius
+    measurements_layer.selected_data = {3, 4, 5}
+    cpt_widget._on_click_radius()
+
+    first_radius = cpt_widget.data[0].radius_px
+    second_radius = cpt_widget.data[1].radius_px
+    assert (
+        first_radius != second_radius
+    ), "The radii of different particles should be different"
+
+    first_rpoints = cpt_widget.data[0].rpoints
+    second_rpoints = cpt_widget.data[1].rpoints
+    assert (
+        first_rpoints != second_rpoints
+    ), "The points for the radii calculation of different particles should be different"
+
+    first_radius = cpt_widget.data[0].radius_cm
+    second_radius = cpt_widget.data[1].radius_cm
+    assert (
+        first_radius != second_radius
+    ), "The radii of different particles should be different"
 
 
 def test_calculate_length_ui(
@@ -185,9 +211,7 @@ def test_calculate_length_ui(
     # click the calculate decay length button
     cpt_widget._on_click_length()
 
-    assert cpt_widget.table.item(
-        0, cpt_widget._get_table_column_index("decay_length_px")
-    )
+    assert cpt_widget.table.item(0, cpt_widget._get_table_column_index("decay_length_px"))
     assert cpt_widget.data[0].decay_length_px == pytest.approx(1, rel=1e-3)
 
 
@@ -212,10 +236,7 @@ def test_calculate_length_fails_with_wrong_number_of_points(
     cpt_widget._on_click_length()
     captured = capsys.readouterr()
 
-    assert (
-        "ERROR: Select two points to calculate the decay length."
-        in captured.out
-    )
+    assert "ERROR: Select two points to calculate the decay length." in captured.out
 
 
 def test_calculate_length_fails_data_out_of_sync(
@@ -223,7 +244,7 @@ def test_calculate_length_fails_data_out_of_sync(
 ):
     # Add images to the viewer
     images = np.random.randint(0, 10, (3, 5, 10, 10), "uint8")
-    cpt_widget.viewer.add_image(images, name="Particle Tracks")
+    cpt_widget.viewer.add_image(images, name=IMAGE_LAYER_NAME)
     cpt_widget.viewer.dims.set_current_step(0, 0)  # move to view 0
     cpt_widget.viewer.dims.set_current_step(1, 0)  # move to event 0
 
@@ -275,3 +296,41 @@ def test_calculate_length_fails_data_out_of_sync(
     assert cpt_widget.data[0].decay_length_px == pytest.approx(
         1, rel=1e-3
     ), "The decay length should be calculated"
+
+
+def test_length_save_preserves_old_data(cpt_widget):
+    """Test saving a new particle length does not mess up the previous one."""
+    measurements_layer = cpt_widget._setup_measurement_layer()
+
+    # Add first particle
+    cpt_widget.particle_decays_menu.setCurrentIndex(1)
+    # Add two points 1 px apart
+    measurements_layer.add([[0, 0, 0, 0], [0, 0, 1, 0]])  # 1px length
+    # Select the first two points in the layer
+    measurements_layer.selected_data = {0, 1}
+    # User clicks the length calculation button
+    cpt_widget._on_click_length()
+
+    # Create a second particle, add points corresponding 2px apart, and click calculate length.
+    cpt_widget.particle_decays_menu.setCurrentIndex(1)
+    measurements_layer.add([[0, 0, 1, 1], [0, 0, 1, 3], [0, 0, 0, 3]])  # 2px length
+    measurements_layer.selected_data = {2, 3}
+    cpt_widget._on_click_length()
+
+    first_decay_length = cpt_widget.data[0].decay_length_px
+    second_decay_length = cpt_widget.data[1].decay_length_px
+    assert (
+        first_decay_length != second_decay_length
+    ), "The decay length of different particles should be different"
+
+    first_decay_length = cpt_widget.data[0].decay_length_cm
+    second_decay_length = cpt_widget.data[1].decay_length_cm
+    assert (
+        first_decay_length != second_decay_length
+    ), "The decay length of different particles should be different"
+
+    first_dpoints = cpt_widget.data[0].dpoints
+    second_dpoints = cpt_widget.data[1].dpoints
+    assert (
+        first_dpoints != second_dpoints
+    ), "The points for the decay length calculation of different particles should be different"
